Refactor issue 3136 regression test
