+++ springdoc-openapi-tests/springdoc-openapi-data-rest-tests/src/test/java/test/org/springdoc/api/v31/app40/GoalInitiativeImpactId.java
@@ -1,0 +1,52 @@
+/*
+ *
+ *  * Copyright 2019-2026 the original author or authors.
+ *  *
+ *  * Licensed under the Apache License, Version 2.0 (the "License");
+ *  * you may not use this file except in compliance with the License.
+ *  * You may obtain a copy of the License at
+ *  *
+ *  *      https://www.apache.org/licenses/LICENSE-2.0
+ *  *
+ *  * Unless required by applicable law or agreed to in writing, software
+ *  * distributed under the License is distributed on an "AS IS" BASIS,
+ *  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  * See the License for the specific language governing permissions and
+ *  * limitations under the License.
+ *
+ */
+
+package test.org.springdoc.api.v31.app40;
+
+import java.io.Serializable;
+
+import jakarta.persistence.Embeddable;
+
+/**
+ * Composite identifier for the join entity used by the issue reproducer.
+ *
+ * @author hej090224
+ */
+@Embeddable
+public class GoalInitiativeImpactId implements Serializable {
+
+	private Long goalId;
+
+	private Long initiativeId;
+
+	public Long getGoalId() {
+		return goalId;
+	}
+
+	public void setGoalId(Long goalId) {
+		this.goalId = goalId;
+	}
+
+	public Long getInitiativeId() {
+		return initiativeId;
+	}
+
+	public void setInitiativeId(Long initiativeId) {
+		this.initiativeId = initiativeId;
+	}
+}
