+++ springdoc-openapi-tests/springdoc-openapi-data-rest-tests/src/test/java/test/org/springdoc/api/v31/app40/InitiativeImpactOnGoal.java
@@ -16,7 +16,7 @@
  *
  */
 
-package test.org.springdoc.api.v31.issue3136;
+package test.org.springdoc.api.v31.app40;
 
 import jakarta.persistence.EmbeddedId;
 import jakarta.persistence.Entity;
