+++ springdoc-openapi-tests/springdoc-openapi-data-rest-tests/src/test/java/test/org/springdoc/api/v31/app40/InitiativeRepository.java
@@ -16,16 +16,16 @@
  *
  */
 
-package test.org.springdoc.api.v31.issue3136;
+package test.org.springdoc.api.v31.app40;
 
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 /**
- * Spring Data REST repository for the issue reproducer aggregate.
+ * Spring Data REST repository for initiatives.
  *
  * @author hej090224
  */
 @RepositoryRestResource
-public interface GoalRepository extends CrudRepository<Goal, Long> {
+public interface InitiativeRepository extends CrudRepository<Initiative, Long> {
 }
