--- springdoc-openapi-tests/springdoc-openapi-data-rest-tests/src/test/java/test/org/springdoc/api/v31/issue3136/Goal.java
@@ -1,59 +1,0 @@
-/*
- *
- *  * Copyright 2019-2026 the original author or authors.
- *  *
- *  * Licensed under the Apache License, Version 2.0 (the "License");
- *  * you may not use this file except in compliance with the License.
- *  * You may obtain a copy of the License at
- *  *
- *  *      https://www.apache.org/licenses/LICENSE-2.0
- *  *
- *  * Unless required by applicable law or agreed to in writing, software
- *  * distributed under the License is distributed on an "AS IS" BASIS,
- *  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  * See the License for the specific language governing permissions and
- *  * limitations under the License.
- *
- */
-
-package test.org.springdoc.api.v31.issue3136;
-
-import java.util.HashSet;
-import java.util.Set;
-
-import jakarta.persistence.CascadeType;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.OneToMany;
-
-/**
- * Aggregate root with a collection of composite-key join entities.
- *
- * @author hej090224
- */
-@Entity
-public class Goal {
-
-	@Id
-	@GeneratedValue(strategy = GenerationType.AUTO)
-	private Long id;
-
-	private String name;
-
-	@OneToMany(mappedBy = "goal", cascade = CascadeType.ALL)
-	private Set<InitiativeImpactOnGoal> impactsByInitiatives = new HashSet<>();
-
-	public Long getId() {
-		return id;
-	}
-
-	public String getName() {
-		return name;
-	}
-
-	public Set<InitiativeImpactOnGoal> getImpactsByInitiatives() {
-		return impactsByInitiatives;
-	}
-}
